@@ -2,55 +2,60 @@
   "name": "audako-powertoys",
   "version": "1.0.0",
   "description": "",
-  "main": "webpack.config.js",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack",
-    "watch": "webpack --watch"
+    "dev": "vite",
+    "clean": "rm -rf dist/build dist/assets",
+    "build": "npm run clean && vite build",
+    "build:debug": "npm run clean && vite build --mode debug",
+    "build:watch": "vite build --watch",
+    "build:debug:watch": "vite build --mode debug --watch",
+    "preview": "vite preview",
+    "update-core": "npm i https://github.com/JacobDeuchert/audako-core-components"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/JacobDeuchert/audako-powertoys.git"
   },
-  "author": "",
-  "license": "ISC",
+  "author": "jacobdeuchert",
+  "license": "MIT",
   "bugs": {
     "url": "https://github.com/JacobDeuchert/audako-powertoys/issues"
   },
   "homepage": "https://github.com/JacobDeuchert/audako-powertoys#readme",
   "dependencies": {
-    "@aspnet/signalr": "^1.0.27",
-    "axios": "^0.22.0",
-    "dayjs": "^1.10.7",
-    "node-sass": "^6.0.1",
-    "reflect-metadata": "^0.1.13",
-    "rxjs": "^7.3.0",
-    "tsyringe": "^4.6.0"
+    "@audako/chat-ui": "file:../audako-mcp-gateway/packages/chat-ui",
+    "@fortawesome/fontawesome-free": "^7.2.0",
+    "@smui/common": "^8.0.3",
+    "@webcomponents/webcomponentsjs": "^2.8.0",
+    "audako-core-components": "github:JacobDeuchert/audako-core-components",
+    "axios": "^1.7.9",
+    "luxon": "^3.7.2",
+    "reflect-metadata": "^0.2.2",
+    "rxjs": "^7.8.1",
+    "svelte-material-ui": "^8.0.3",
+    "tsyringe": "^4.8.0"
   },
   "devDependencies": {
-    "@smui/button": "^4.2.0",
-    "@smui/checkbox": "^4.2.0",
-    "@smui/dialog": "^4.2.0",
-    "@smui/icon-button": "^4.2.0",
-    "@smui/list": "^4.2.0",
-    "@smui/menu": "^4.2.0",
-    "@smui/paper": "^4.2.0",
-    "@smui/ripple": "^4.2.0",
-    "@smui/segmented-button": "^4.2.0",
-    "@smui/snackbar": "^4.2.0",
-    "@smui/textfield": "^4.2.0",
-    "@smui/tooltip": "^4.2.0",
-    "@types/chrome": "0.0.158",
-    "copy-webpack-plugin": "^9.0.1",
-    "css-loader": "^6.3.0",
-    "mini-css-extract-plugin": "^2.3.0",
-    "svelte": "^3.43.0",
-    "svelte-loader": "^3.1.2",
-    "svelte-material-ui": "^4.2.0",
-    "svelte-preprocess": "^4.9.5",
-    "ts-loader": "^9.2.6",
-    "typescript": "^4.4.3",
-    "webpack": "^5.54.0",
-    "webpack-cli": "^4.8.0"
+    "@smui/button": "^8.0.3",
+    "@smui/checkbox": "^8.0.3",
+    "@smui/dialog": "^8.0.3",
+    "@smui/icon-button": "^8.0.3",
+    "@smui/list": "^8.0.3",
+    "@smui/menu": "^8.0.3",
+    "@smui/paper": "^8.0.3",
+    "@smui/ripple": "^8.0.3",
+    "@smui/segmented-button": "^8.0.3",
+    "@smui/snackbar": "^8.0.3",
+    "@smui/textfield": "^8.0.3",
+    "@smui/tooltip": "^8.0.3",
+    "@sveltejs/vite-plugin-svelte": "^5.0.2",
+    "@types/chrome": "^0.0.277",
+    "sass": "^1.83.4",
+    "svelte": "^5.50.0",
+    "typescript": "^5.7.2",
+    "vite": "^6.0.5",
+    "vite-plugin-static-copy": "^2.1.0"
   }
 }
