@@ -3,51 +3,45 @@
   import IconButton from '@smui/icon-button';
   import Textfield from '@smui/textfield';
   import List, { Item, Separator, Text, Graphic } from '@smui/list';
-  import { container } from 'tsyringe';
-  import { HttpService } from '../../../services/http.service';
-  import { SignalRService } from '../../../services/signalR.service';
-  import { SvelteComponent } from 'svelte';
   import { debounceTime, filter, first, map, Subject } from 'rxjs';
   import { catchPromise } from '../../../utils/promise-utils';
   import { SearchService } from './search.service';
-  import { CategorizedSearchResults, CategorySearchResult, SearchCategory, SearchResult } from './search-results';
-import { DomUtils } from '../../../utils/dom-utils';
-import App from '../../App.svelte';
-import AppSwitch from '../AppSwitch/AppSwitch.svelte';
+  import type { CategorizedSearchResults, CategorySearchResult, SearchCategory, SearchResult } from './search-results';
+  import { DomUtils } from '../../../utils/dom-utils';
 
-  interface DisplayedCategoySearchResults {
+  interface DisplayedCategorySearchResults {
     category: SearchCategory;
     displayedResults: SearchResult[];
     totalResults: SearchResult[];
   }
+
   
-  const httpService = container.resolve(HttpService);
-  const signalRService = container.resolve(SignalRService);
+  const searchService = new SearchService();
 
-  const searchService = new SearchService(httpService, signalRService);
+  let displayedResults = $state<DisplayedCategorySearchResults[]>([]);
 
-  let displayedResults: DisplayedCategoySearchResults[] = [];
+  let focusedCategory = $state<SearchCategory | null>(null);
 
-  let focusedCategory: SearchCategory = null;
+  let searchOpen = $state<boolean>(false);
+  let searchString = $state<string>('');
 
-  let searchOpen: boolean = false;
-  let searchString: string = '';
+  let encounteredError = $state<boolean>(false);
 
-  let encounteredError: boolean = false;
-
-  let inputComponent: SvelteComponent;
-  let listComponent: SvelteComponent;
+  let inputComponent = $state<any>(null);
+  let listComponent: any;
 
   let searchChanged$: Subject<string> = new Subject<string>();
 
-  $: searchChanged$.next(searchString);
+  $effect(() => {
+    searchChanged$.next(searchString);
+  });
 
   searchChanged$
     .asObservable()
     .pipe(
       debounceTime(300),
       filter((x) => typeof x === 'string' && x?.length > 1),
-      map((search) => search.trim().toLowerCase())
+      map((search) => search.trim().toLowerCase()),
     )
     .subscribe(async (searchString) => {
       console.log('Searching', searchString);
@@ -64,15 +58,18 @@ import AppSwitch from '../AppSwitch/AppSwitch.svelte';
       console.log('Search results', displayedResults);
     });
 
+console.log('Search component loaded', window);
+
   window.addEventListener('keydown', (event: KeyboardEvent) => {
     if (event.ctrlKey && event.code === 'Space') {
       searchOpen = true;
-      focuSearchInput(true);
+      console.log('Opening search');
+      focusSearchInput(true);
     }
 
     const excludedFocusKeys = ['ArrowUp', 'ArrowDown', 'ArrowLeft', 'ArrowRight', 'Tab', 'Enter'];
     if (searchOpen && !event.ctrlKey && !event.shiftKey && !event.altKey && !excludedFocusKeys.includes(event.code)) {
-      focuSearchInput();
+      focusSearchInput();
     }
   });
 
@@ -81,7 +78,8 @@ import AppSwitch from '../AppSwitch/AppSwitch.svelte';
     const listElement = listComponent.getElement();
 
     if (event.code === 'ArrowDown') {
-      const activeElement = document.activeElement;
+      const rootNode = inputElement?.getRootNode() as Document | ShadowRoot;
+      const activeElement = rootNode?.activeElement ?? document.activeElement;
       const firstListElement = listElement.getElementsByTagName('li').item(0);
       if (firstListElement && activeElement === inputElement) {
         firstListElement.focus();
@@ -89,50 +87,54 @@ import AppSwitch from '../AppSwitch/AppSwitch.svelte';
     }
   });
 
-  function getDisplayedResults(searchResults: CategorizedSearchResults): DisplayedCategoySearchResults[]{
+  function getDisplayedResults(searchResults: CategorizedSearchResults): DisplayedCategorySearchResults[] {
     // display 4 of each category and 8 if only 1 category has results
     // display max 16 results
 
-    
     const categoriesWithResults = searchResults.filter((x) => x.results.length > 0);
 
     const resultsPerCategory = categoriesWithResults.length > 1 ? 4 : 8;
 
-    return searchResults.map((x) => (
-      {
-        displayedResults: x.results.slice(0, resultsPerCategory), 
-        category: x.category,
-        totalResults: x.results
-      }));
+    return searchResults.map((x) => ({
+      displayedResults: x.results.slice(0, resultsPerCategory),
+      category: x.category,
+      totalResults: x.results,
+    }));
   }
 
   function showAllResults(category: SearchCategory) {
     const result = displayedResults.find((x) => x.category === category);
-    displayedResults = [{
-      category,
-      displayedResults: result.totalResults,
-      totalResults: result.totalResults
-    }];
+    displayedResults = [
+      {
+        category,
+        displayedResults: result.totalResults,
+        totalResults: result.totalResults,
+      },
+    ];
 
     const listElement = listComponent.getElement();
     console.log(listElement);
-    DomUtils.watchForDomChanges(listElement).pipe(first()).subscribe((changes) => {
-      console.log(changes);
-      const firstListElement = listElement.getElementsByTagName('li').item(result.displayedResults.length);
-      if (firstListElement) {
-        firstListElement.focus();
-      }
-    });
+    DomUtils.watchForDomChanges(listElement)
+      .pipe(first())
+      .subscribe((changes) => {
+        console.log(changes);
+        const firstListElement = listElement.getElementsByTagName('li').item(result.displayedResults.length);
+        if (firstListElement) {
+          firstListElement.focus();
+        }
+      });
   }
 
-
-  function focuSearchInput(select: boolean = false) {
-    const textField = document.getElementsByClassName('search-text-field').item(0);
+  function focusSearchInput(select: boolean = false) {
+    console.log('Focusing search input');
+    const textField = inputComponent?.getElement();
 
     if (textField) {
       const input = textField.getElementsByTagName('input').item(0);
+      const rootNode = input?.getRootNode() as Document | ShadowRoot;
+      const activeElement = rootNode?.activeElement ?? document.activeElement;
 
-      if (document.activeElement !== input) {
+      if (activeElement !== input) {
         input.focus();
         if (select) {
           input.select();
@@ -142,7 +144,7 @@ import AppSwitch from '../AppSwitch/AppSwitch.svelte';
   }
 </script>
 
-<Dialog class="search-dialog" bind:open={searchOpen}>
+<Dialog bind:open={searchOpen}>
   <Content style="padding: 12px" class="search-dialog-content">
     <Textfield
       class="search-text-field"
@@ -152,17 +154,17 @@ import AppSwitch from '../AppSwitch/AppSwitch.svelte';
       variant="outlined"
       type="text"
       label="Search"
-      bind:input={inputComponent}
+      bind:this={inputComponent}
     />
 
     <div class="result-list-container" style="margin-top: 8px;">
-      <List bind:this={listComponent} class="result-list" bind>
+      <List bind:this={listComponent} class="result-list">
         {#each displayedResults as categoryResults, categoryIndex}
           <div style="font-size: 13px; font-weight: bold; opacity: 0.7; margin-top: -4px">
             {categoryResults.category}
           </div>
           {#each categoryResults.displayedResults as result, resultIndex}
-            <Item class="list-item" on:SMUI:action={() => result.defaultAction()}>
+            <Item class="list-item" onclick={() => result.defaultAction()}>
               <Graphic class="search-icon {result.icon}" />
               <Text>
                 {result.title}
@@ -174,28 +176,27 @@ import AppSwitch from '../AppSwitch/AppSwitch.svelte';
                 </Text>
               {/if}
 
-              
               {#if result.infoComponent}
-              <div class="info-component">
-                <svelte:component this={result.infoComponent.component} {...result.infoComponent.props} />
-              </div>
+                {@const InfoComp = result.infoComponent.component}
+                <div class="info-component">
+                  <InfoComp {...result.infoComponent.props} />
+                </div>
               {/if}
 
               {#if result.extraActions?.length > 0}
-                <div class="action-buttons" on:click={(event) => event.stopPropagation()}>
+                <div class="action-buttons" onclick={(event) => event.stopPropagation()}>
                   {#each result.extraActions as action}
-                    <IconButton on:click={() => action.onClick()}>
-                      <i class={action.icon} />
+                    <IconButton onclick={() => action.onClick()}>
+                      <i class={action.icon}></i>
                     </IconButton>
                   {/each}
                 </div>
               {/if}
-
             </Item>
           {/each}
           <!-- Show "View More" when displayed items are not equal the search results  -->
           {#if categoryResults.displayedResults.length < categoryResults.totalResults.length}
-            <Item class="list-item" on:SMUI:action={() => showAllResults(categoryResults.category)}>
+            <Item class="list-item" onclick={() => showAllResults(categoryResults.category)}>
               <Text>
                 View All Results (+{categoryResults.totalResults.length - categoryResults.displayedResults.length})
               </Text>
@@ -211,13 +212,12 @@ import AppSwitch from '../AppSwitch/AppSwitch.svelte';
 </Dialog>
 
 <style>
-
-  .action-buttons {
+  /* .action-buttons {
     visibility: hidden;
     margin-left: auto;
     display: flex;
   }
-  
+
   .info-component {
     margin-left: auto;
   }
@@ -231,7 +231,7 @@ import AppSwitch from '../AppSwitch/AppSwitch.svelte';
     max-height: 100%;
   }
 
-  :global(.search-dialog){
+  :global(.search-dialog) {
     max-height: 50vh;
     overflow: hidden;
   }
@@ -241,7 +241,7 @@ import AppSwitch from '../AppSwitch/AppSwitch.svelte';
     margin-top: 30px;
   }
 
-  :global(.search-dialog .mdc-dialog__surface){
+  :global(.search-dialog .mdc-dialog__surface) {
     width: 40vw;
   }
 
@@ -257,7 +257,7 @@ import AppSwitch from '../AppSwitch/AppSwitch.svelte';
     }
   }
 
-  :global(.search-dialog .mdc-dialog__content){
+  :global(.search-dialog .mdc-dialog__content) {
     overflow: hidden;
     display: flex;
     flex-direction: column;
@@ -303,4 +303,5 @@ import AppSwitch from '../AppSwitch/AppSwitch.svelte';
   :global(.list-item:focus-within .action-buttons) {
     visibility: visible;
   }
+    */
 </style>
