@@ -1,31 +1,18 @@
+import { EntityType, Group } from 'audako-core-components';
 import { AppIcons, AudakoApp } from '../../../../models/audako-apps';
-import { EntityType } from '../../../../models/configuration-entity';
-import { Group } from '../../../../models/group';
-import { HttpService } from '../../../../services/http.service';
 import { UrlUtils } from '../../../../utils/url-utils';
-import { IndexedTenant } from '../search-requisites';
 import { ResultAction, SearchResult } from '../search-results';
-import { TenantIndexer } from '../tenant-indexer';
 import { SearchQuery } from './search-query';
 
 export class GroupQuery extends SearchQuery {
 
   private readonly DEFAULT_ICON = 'fas fa-folder';
 
-  constructor(tenantIndexer: TenantIndexer, httpService: HttpService) {
-    super(tenantIndexer, httpService);
+  constructor() {
+    super();
   }
 
   public async query(queryString: string, tenantRestriction?: string): Promise<SearchResult[]> {
-
-    const indexedTenants = await this.tenantIndexer.getIndexedTenants();
-    
-    const defaultAction = (tenant: IndexedTenant, group: Group) =>
-      UrlUtils.openApp(AudakoApp.Configuration, tenant.Id, group.Id);
-
-    const tooltip = (group: Group) => {
-      return this.httpService.resolvePathName(group.Path);
-    };
 
     const groupProjection = { Name: 1, Path: 1, Type: 1, IsEntryPoint: 1 };
     const matchedGroups = await this.requestConfigurationEntities<Group>(
@@ -35,31 +22,29 @@ export class GroupQuery extends SearchQuery {
       groupProjection
     );
 
-    return matchedGroups.map((group) => {
-      const tenant = this.getTenantForEntity(indexedTenants, group);
+    return Promise.all(
+      matchedGroups.map(async (group) => {
+        const tenant = await this.getTenantForEntity(group);
+        const tenantId = tenant?.Id ?? tenant?.Root ?? group.Path?.[0] ?? group.Id;
 
-      const actionButtons: ResultAction[] = [
-        {
-          onClick: () => UrlUtils.openApp(AudakoApp.Commissioning, tenant.Id, group.Id),
-          icon: AppIcons.Commissioning,
-        },
-      ];
+    const actionButtons: ResultAction[] = [];
 
-      if (group.IsEntryPoint) {
-        actionButtons.unshift({
-          onClick: () => UrlUtils.openApp(AudakoApp.Dashboard, tenant.Id, group.Id),
-          icon: AppIcons.Dashboard,
-        });
-      }
+    if (group.IsEntryPoint) {
+      actionButtons.push({
+        onClick: () => UrlUtils.openApp(AudakoApp.Dashboard, tenantId, group.Id),
+        icon: AppIcons.Dashboard,
+      });
+    }
 
-      return {
-        title: group.Name.Value,
-        infoText: tenant?.Name,
-        defaultAction: () => defaultAction(tenant, group),
-        extraActions: actionButtons,
-        icon: this.DEFAULT_ICON,
-        tooltip: () => tooltip(group),
-      };
-    });
+        return {
+          title: group.Name.Value,
+          infoText: tenant?.Name,
+          defaultAction: () => UrlUtils.openApp(AudakoApp.Configuration, tenantId, group.Id),
+          extraActions: actionButtons,
+          icon: this.DEFAULT_ICON,
+          tooltip: () => this.entityNameService.resolvePathName(group.Path ?? []),
+        };
+      })
+    );
   }
 }
