@@ -1,24 +1,31 @@
-import { ConfigurationEntity, EntityType } from '../../../../models/configuration-entity';
-import { HttpService } from '../../../../services/http.service';
-import { IndexedTenant } from '../search-requisites';
+import { ConfigurationEntity, EntityHttpService, EntityNameService, EntityType, TenantHttpService, TenantView } from 'audako-core-components';
+import { container } from 'tsyringe';
 import { SearchResult } from '../search-results';
-import { TenantIndexer } from '../tenant-indexer';
+import { resolveService } from 'audako-core-components';
 
 export abstract class SearchQuery {
-  constructor(protected tenantIndexer: TenantIndexer, protected httpService: HttpService) {}
+  protected tenantHttpService: TenantHttpService;
+  protected entityHttpService: EntityHttpService;
+  protected entityNameService: EntityNameService;
+  private tenantById: Map<string, TenantView>;
+  private tenantByEntityId: Map<string, TenantView>;
+
+  constructor() {
+    this.tenantHttpService = resolveService(TenantHttpService);
+    this.entityHttpService = resolveService(EntityHttpService);
+    this.entityNameService = resolveService(EntityNameService);
+    this.tenantById = new Map();
+    this.tenantByEntityId = new Map();
+  }
 
   public abstract query(queryString: string, tenantRestriction?: string): Promise<SearchResult[]>;
-
 
   protected async requestConfigurationEntities<T extends ConfigurationEntity>(
     entityType: EntityType,
     searchString: string,
-    tenantRestriction: string,
+    tenantRestriction?: string,
     projection?: { [p in keyof T]?: number }
   ): Promise<T[]> {
-
-    const indexedTenants = await this.tenantIndexer.getIndexedTenants();
-
     const searchParts = searchString.split(' ');
     const fullMatchFilter = { 'Name.Value': { $regex: searchString, $options: 'i' } };
     const partMatchFilter = { $and: searchParts.map((x) => ({ 'Name.Value': { $regex: x, $options: 'i' } })) };
@@ -27,10 +34,11 @@ export abstract class SearchQuery {
 
 
     if (tenantRestriction && tenantRestriction.length > 0) {
-      const tenant = indexedTenants.find((t) => t.Id === tenantRestriction);
+      const tenant = await this.getTenantById(tenantRestriction);
 
       if (tenant) {
-        const pathFilter = { $or: [{ Path: tenant.Root }, { Id: tenant.Root }] };
+        const rootId = tenant.Root ?? tenant.Id;
+        const pathFilter = { $or: [{ Path: rootId }, { Id: rootId }] };
         filter = { $and: [filter, pathFilter] };
       }
     }
@@ -40,19 +48,69 @@ export abstract class SearchQuery {
 
     projection = projection || { Id: 1, Name: 1, Path: 1, GroupId: 1 };
 
-    const result = await this.httpService.queryConfiguration<T>(entityType, filter, paging, projection);
+    const result = await this.entityHttpService.queryConfiguration<T>(entityType, filter, paging, projection);
+    const data = result?.data ?? [];
 
-    return result.sort((a, b) => a.Path.length - b.Path.length) as T[];
+    return data.sort((a, b) => (a.Path?.length ?? 0) - (b.Path?.length ?? 0)) as T[];
   }
 
 
-  public getTenantForEntity(indexedTenants: IndexedTenant[], entity: ConfigurationEntity): IndexedTenant {
-    let rootPathEntry = entity.Path[0];
-
-    if (!rootPathEntry || rootPathEntry.length === 0) {
-      rootPathEntry = entity.Id;
+  public async getTenantForEntity(entity: ConfigurationEntity): Promise<TenantView | null> {
+    if (!entity?.Id) {
+      return null;
     }
 
-    return indexedTenants?.find(t => t.Root === rootPathEntry);
+    const cachedTenant = this.tenantByEntityId.get(entity.Id);
+    if (cachedTenant) {
+      return cachedTenant;
+    }
+
+    try {
+      const tenant = await this.tenantHttpService.getTenantViewForEntityId(entity.Id);
+      if (tenant) {
+        this.cacheTenant(tenant);
+        this.tenantByEntityId.set(entity.Id, tenant);
+      }
+      return tenant ?? null;
+    } catch {
+      return null;
+    }
   }
+
+  protected async getTenantById(tenantId: string): Promise<TenantView | null> {
+    if (!tenantId) {
+      return null;
+    }
+
+    const cachedTenant = this.tenantById.get(tenantId);
+    if (cachedTenant) {
+      return cachedTenant;
+    }
+
+    try {
+      const tenant = await this.tenantHttpService.getTenantViewById(tenantId);
+      if (tenant) {
+        this.cacheTenant(tenant);
+      }
+      return tenant ?? null;
+    } catch {
+      return null;
+    }
+  }
+
+  private cacheTenant(tenant: TenantView): void {
+    if (!tenant) {
+      return;
+    }
+
+    this.tenantById.set(tenant.Id, tenant);
+    if (tenant.Root) {
+      this.tenantById.set(tenant.Root, tenant);
+    }
+
+    if (tenant.Id === tenant.Root) {
+      this.tenantByEntityId.set(tenant.Id, tenant);
+    }
+  }
+
 }
